pyproject.toml
[project]
name = "gwaphics-dirt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame>=2.6.1",
    "pyglm>=2.8.3",
]

[project.scripts]
gra = "gra.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
